Modularize variables
